spec: let -tests subpackage own ostree-trivial-httpd
We were missing another exclude for ostree-trivial-httpd. This was
causing it to be shipped in the main ostree package instead, which in
turned pulled libsoup back into AH.

I kept the globby grab for everything else in
`%{_libexecdir}/libostree}`, even though all the files we currently
install in there are picked up by either the -grub2 or -tests
subpackages.