Disable HTTP2 for now
I'm beating Colin to the punch here and making use of the new
`--disable-http2`. What pushed me over the edge was the fact that I
literally could not upgrade my Atomic Workstation with a simple
`rpm-ostree upgrade` without hitting:

https://github.com/ostreedev/ostree/issues/1362

I originally thought this only occurred in low-bandwidth environments,
when in fact it still happens even after upgrading my Internet bundle
and can actually be reproduced very easily in environments where
bandwidth is not an issue. E.g. this currently hangs in the latest
Fedora 27 Atomic Host provisioned in the cloud:

    ostree remote add --no-gpg-verify faw27 https://dl.fedoraproject.org/ostree/27/
    ostree pull faw27 fedora/27/x86_64/workstation

Until we investigate deeper, let's just play it safe and disable HTTP2.
This should also fix the HTTP2 framing layer people have been hitting.

Let's get this in to confirm it doesn't break the rdgo streams, and then
we can backport the patch for f27 to make sure it gets into the next
TWR?