spec: Add requires from subpackages to libs
Got rpmdiff complaining that the `ostree` and `ostree-tests` packages do
not have an explicit `Requires` on the libraries. Now, symbol versioning
assures us that e.g. the -libs version that we pull for a given `ostree`
package will have the symbols we need, but it doesn't guarantee that
they'll be the exact same version. Let's add the explicit `Requires` to
make this clear since in practice we don't want to have to think about
such mismatches.

I've also taken the opportunity to normalize on the usage of `_isa` and
`epoch` in all the `Requires`.