spec: Default to not shipping tests on RHEL > 7
We weren't shipping it in RHEL7 either. It also brings in a dependency
on Python, which is tricky there. (And anyway
`gnome-desktop-testing-runner` isn't even packaged there).

Make it a bcond though so one can still build it if one desires. (And
symmetrically, to *not* build it on Fedora if one desires).